chore: moves bls back to agent
