+++ packages/agent/src/index.ts
@@ -9,6 +9,7 @@ export * from './agent/http/types';
 export * from './canisters/asset';
 export * from './canisters/management';
 export * from './request_id';
+export * from './utils/bls';
 export * as polling from './polling';
 
 import { Agent, HttpAgent } from './agent';
