+++ packages/agent/tsconfig.json
@@ -21,5 +21,5 @@
     "target": "es2017",
     "tsBuildInfoFile": "./build_info.json"
   },
-  "include": ["types/*", "src/**/*.ts"]
+  "include": ["types/*", "src/**/*.ts", "../candid/src/utils/bls.test.ts"]
 }
