+++ packages/candid/src/index.ts
@@ -1,7 +1,6 @@
 export * from './candid-ui';
 export * from './candid-core';
 export * as IDL from './idl';
-export * from './utils/bls';
 export * from './utils/hash';
 export * from './utils/leb128';
 export * from './types';
